# Improve performance of binary search helpers