Avoid pure-python element comparisons since numpy is rather
slow at that. Try to do that in native math when possible.

Patched:
```
> python -m timeit -s 'import numpy ; from sharedbuffers.mapped_struct import sorted_contains ; a = numpy.arange(1000)' 'sorted_contains(a, 342)'
1000000 loops, best of 3: 0.375 usec per loop

> python -m timeit -s 'import numpy ; from sharedbuffers.mapped_struct import sorted_contains ; a = numpy.arange(1000, dtype=numpy.uint32)' 'sorted_contains(a, -342)'
1000000 loops, best of 3: 1.88 usec per loop

> python -m timeit -s 'import numpy ; from sharedbuffers.mapped_struct import sorted_contains ; a = numpy.arange(1000, dtype=numpy.uint32)' 'sorted_contains(a, 1<<40)'
1000000 loops, best of 3: 0.354 usec per loop
```

Unpatched:
``` pycon
> python -m timeit -s 'import numpy ; from sharedbuffers.mapped_struct import sorted_contains ; a = numpy.arange(1000)' 'sorted_contains(a, 342)'
1000000 loops, best of 3: 0.726 usec per loop

> python -m timeit -s 'import numpy ; from sharedbuffers.mapped_struct import sorted_contains ; a = numpy.arange(1000, dtype=numpy.uint32)' 'sorted_contains(a, -342)'
100000 loops, best of 3: 2.1 usec per loop

> python -m timeit -s 'import numpy ; from sharedbuffers.mapped_struct import sorted_contains ; a = numpy.arange(1000, dtype=numpy.uint32)' 'sorted_contains(a, 1<<40)'
100000 loops, best of 3: 2.14 usec per loop
```

As you can see, some cases, especially the most common cases, improve considerably